requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.25.0
accelerate>=0.20.0
datasets>=2.14.0
Pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface_hub>=0.16.0
safetensors>=0.3.1
timm>=0.9.0
fairscale>=0.4.13
bitsandbytes>=0.41.0
wandb>=0.15.0
xformers>=0.0.20
peft>=0.5.0
scipy>=1.10.0
scikit-learn>=1.2.0
